requirements.txt
openai>=1.3.5
python-dotenv
beautifulsoup4
requests
selenium
couchbase>=4.1.6
